Include subdirs in go refactor mapping + replacement pre-selecting removed.
The mapping using gofmt to rename an item on <F6> doesn't automatically search
recursively. Adding **/*.go changes this. In other words:

- !gofmt -r 'old -> new' -w *.go
+ !gofmt -r 'old -> new' -w *.go **/*.go

Previously the replacement word was pre-selected so it could be direclty
replaced. But more often then not I want to rename only part of the word, so I
ended up almost always deselecting first.