fix (re)runn tests using filetypes and dispatch
Rerunning tests from a window with another filetype, for example the
quickfix window, now uses the saved filetype from the last run to simply
rerun the last test correctly.

Add a fallback for rerunning tests to simply run :Dispatch if any
dispatch command is available. This allows for rerunning to fallback to
Dispatch. Now a custom :ReDispatch command can be rerun using the same
mapping.